Update Splash Screen Code
Splash screen appears as an introductory screen of an application. It also signifies that you have to wait for a few seconds before landing on the actual screen of the application.

in my code, i created a JLabel and i added our logo to it using ImageIcon, then i added this label to my JWindow. after that, i used Thread.Sleep(long millies) to wait for three seconds before launching our app.